+++ pint/facets/nonmultiplicative/__init__.py
@@ -1,0 +1,18 @@
+"""
+    pint.facets.group
+    ~~~~~~~~~~~~~~~~~
+
+    Adds pint the capability to group units.
+
+    :copyright: 2022 by Pint Authors, see AUTHORS for more details.
+    :license: BSD, see LICENSE for more details.
+"""
+
+# This import register LogarithmicConverter and OffsetConverter to be usable
+# (via subclassing)
+from .definitions import LogarithmicConverter, OffsetConverter  # noqa: F401
+from .registry import NonMultiplicativeRegistry
+
+__all__ = [
+    NonMultiplicativeRegistry,
+]
